Industry Analysis v1
Sorry guys for all the commits, I have the feeling I still need to learn a lot about Git. I tried to write this in Atom and then commit it, but apparently it doesn't do it automatically in .md format so that's why it went wrong. Afterwards I could not merge anymore because of a conflict between my local folder and the online repository.
Still, I think there is something wrong with this file as it starts with a red line and then doctype html.